+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Info can only be retrieved by calling `getInfo`
 * Removed third parameter from _insertFeature as it was not being used for the id
 * Casting values to decimal before calling statistic functions
+* Use geometry field for spatial queries to improve speed
 
 ## [1.3.1] - 2015-09-15 
 ### Fixed